+++ main.py
@@ -1,5 +1,6 @@
 import Common.url as Url
 from OpticalFlow.opticalflow_sparse import OpticalFlowSparse
+from OpticalFlow.opticalflow_dense import OpticalFlowDense
 
 DENSE = "Dense"
 SPARSE = "Sparse"
@@ -13,11 +14,16 @@ class App:
                                            height_cam=height_cam,
                                            width_cam=width_cam)
 
-        self.op_dense = None
+        self.op_dense = OpticalFlowDense(video_url=video_url,
+                                         height_cam=height_cam,
+                                         width_cam=width_cam)
 
     def run(self):
         if self.type_op == SPARSE:
             self.op_sparse.run()
+
+        else:
+            self.op_dense.run()
 
 
 if __name__ == "__main__":
